perf(runtime): ScrStr's header is 12 bytes, and the win is a whole allocator bucket per string
Peak RSS is 100% what the container RETAINS: on the messaging bench's
SEND 1:1 the live ScrStr population at the peak is 1,601,108 strings
holding 140,886,736 B, which is 78.8% of the whole process's 174.5 MiB
peak. ScrStr is rank 1 there, and this is the term.

WHAT WAS MEASURED FIRST. The allocator's bucket function on
x86_64-windows-gnu with zig 0.16.0's mingw CRT is roundup16(request + 8)
- an 8-byte block header and a 16-byte grain - established by allocating
300,000 same-size blocks and reading the modal pointer stride together
with the working-set delta per block, which agree to 0.8 B. Requests of
40/44/48/56/64/72/80/88/89/96/104/112/128 cost
48/64/64/64/80/80/96/96/112/112/112/128/144.

So the saving is NOT eight bytes a string, or twelve. Reducing the
request moves a whole 16-byte bucket on the capacities that cross a
boundary and NOTHING on the rest, and which of the two you get depends
entirely on the population. Over a uniform capacity 0..1000 a 12-byte
header saves a mean of 8.06 B; over THIS bench's real distribution it
saves 16 on 99.9% of the strings.

TWELVE, NOT SIXTEEN. `size_t rc; uint32_t len; uint32_t cap;` is the
obvious narrowing and is worth exactly half of this on RECV 1:1, which
retains 450,000 strings of capacity 10: 24 and 16 both cost 48 bytes
there and only 12 reaches 32. Sixteen is a local minimum that looks like
the answer.

TWELVE, NOT EIGHT. Dropping `cap` and deriving it from the block's size
class is worth a further 0.02% on this population - every capacity a
12-byte header moves, an 8-byte one moves to the same bucket - and it
costs the in-place concat arm its capacity test. Measured, then not
done.

WHAT MOVED, live string bytes at the peak, measured on both sides:

  SEND 1:1     140,886,736 -> 115,342,912   -18.1%
  RECV 1:1      88,144,064 ->  70,612,384   -19.9%
  SEND group    16,123,920 ->  12,921,456   -19.9%
  RECV group    38,414,864 ->  32,056,160   -16.6%

against a projection the reader makes from the BASE population's row
table alone, which the measurement matches to a tenth of a point.

THE HAZARD, AND WHAT MAKES IT LOUD. Four separate places encode this
layout and no compiler checks any of them against the others: the type;
five immortal statics inside scr_string.c; the C backend's literal
table; and the LLVM backend's own `%ScrStr` plus seven raw byte offsets
into it. An ScrStr has a flexible array member, so no literal can be
declared with the type - every one is an anonymous struct cast to
`ScrStr *`, and a width that changed in one place and not the other is a
type pun that builds clean and reads every string in the program at the
wrong offset.

The shape is now scr_runtime.h's SCR_STR_LIT macro, used by the runtime's
five statics, the runtime's C tests and the C emitter alike, with static
asserts tying it to the real type. Spelling the macro back to `size_t`
and building is 6 errors naming the field that moved, where before it was
silence. The LLVM backend has no such tie and cannot have one, which is
what tests/corpus/6301 is for.

Every length that reaches a 32-bit field passes scr_str_size_check, which
TRAPS rather than truncating: a 5 GiB string used to be an allocation
failure and must not become a string that quietly reports the wrong
length. scr_str_concat's sums are cast on an OPERAND rather than on the
result, because `a->len + b->len` over two uint32 fields is evaluated in
unsigned int and wraps BEFORE any guard on the sum could see it.

`rc` narrows too, so the immortal sentinel is SCR_STR_IMMORTAL
(UINT32_MAX) rather than SIZE_MAX. Every other immortal in this runtime
- ScrArr, ScrUnion, ScrClosure, ScrRegex, ScrDyn - keeps a size_t rc and
SIZE_MAX, so only the string sites move. A missed one compiles clean and
is always false; the sweep for them is clang's own
-Wtautological-constant-out-of-range-compare over every runtime source
and every runtime C test, and it found one that the reader would not
have: packages/runtime/test/test_array.c's "immortal element stays
immortal" check.

-Wshorten-64-to-32 over every runtime source reports the SAME eight sites
before and after, none of them a string.

The corpus programs are 6300 (every static string surface: length and
indexing over non-BMP input, slicing, the three concat paths, sort order,
identity against equality, template literals, JSON escaping, regex,
Buffer round-trips, strings as Map and object keys with integer-like keys
first, the immortal statics, and a 1..300 sweep straddling the pool
ceiling - which MOVES from cap 231 to cap 243 with the header) and 6301,
which exists because 6300 emits ZERO inline ScrStr field accesses on the
LLVM backend: reverting all seven of that backend's raw ->data offsets
left 6300 byte-exact. 6301 is `@dynamic`, emits 11 of them, and the same
sabotage makes it diverge on 23 lines.

Both are byte-exact against Node v25.9.0 on both backends, and byte-exact
on the BASE runtime too, so neither tests behaviour this commit
introduced.

THREE ANCHORS FIRED, all three correctly, none a behaviour regression.

tests/harness/size-class.ts's recorded pair is two-sided, and both classes
SHRANK: static hello-world 664,064 -> 656,384 (-7,680) and the regex class
805,376 -> 798,720 (-6,656), measured A/B on two worktrees at the same
commit. They move by different amounts, which is why that file says twice
never to derive one from the other. Base drift since the last calibration
was 0 bytes static and +512 regex, so the previous figures were honest.

And the recalibration corrected an error: the bytes are NOT the emitted
literal table. The messaging bench's TU declares 64 literals, worth 768
bytes, and that binary shrank by the same 7,680 as a hello-world with
three. A shrink identical across two programs of very different literal
counts is a constant of the always-linked runtime - scr_ascii1's 128
interned one-character strings go 32 bytes each to 16 (-2,048 alone), the
other four immortals -16 each, and the rest is code, every s->len and
s->cap read now being a 32-bit load with a shorter encoding. The two
figures being exactly equal is PE alignment: 7,680 is 15 x 512.

The MAX ceilings come DOWN with the recorded pair, by the rule that file
adopted (CLASS_MAX = CLASS_RECORDED + 2 x SIZE_DRIFT_PAGE): 664,576 and
806,912. It is the first entry in that file where a measurement moved
downward, and a ceiling left 8 KB above a binary that no longer weighs
that much is the loose canary its own entries keep warning about.

order-parity.json gains exactly the two new corpus programs, recorded with
SCRIPTC_UPDATE_BASELINES=1 (additive, and it cannot overwrite a live
divergence). Verified key by key rather than by trusting the mode name:
1,854 -> 1,856 entries, two added, none removed, none changed.

+++ packages/runtime/src/scr_regex.c
@@ -1158,7 +1158,7 @@ ScrStr *scr_regexp_escape(ScrStr *s) {
     first = false;
     i += adv;
   }
-  out->len = out_len;
+  out->len = (uint32_t)out_len;
   out->data[out_len] = '\0';
   return out;
 }
@@ -1222,7 +1222,7 @@ static ScrDyn *scr_regex_dynh_to_string(const ScrRegex *re) {
   w += re->source->len;
   *w++ = '/';
   memcpy(w, re->flags->data, re->flags->len);
-  out->len = len;
+  out->len = (uint32_t)len;
   out->data[len] = '\0';
   ScrDyn *d = scr_dyn_new_str(out);
   scr_str_release(out);
